feat(@spartacus/storefront): render address book page using configurable templates (#1081)
Closes #1080
