--- projects/storefrontlib/src/lib/ui/layout/address-book-page-layout/address-book-page-layout.component.html
@@ -1,1 +1,0 @@
-<cx-address-book></cx-address-book>
