--- projects/storefrontlib/src/lib/ui/layout/address-book-page-layout/address-book-page-layout.component.ts
@@ -1,12 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'cx-address-book-page-layout',
-  templateUrl: './address-book-page-layout.component.html',
-  styleUrls: []
-})
-export class AddressBookPageLayoutComponent implements OnInit {
-  constructor() {}
-
-  ngOnInit() {}
-}
