+++ projects/storefrontlib/src/lib/ui/layout/layout.module.ts
@@ -8,7 +8,6 @@ import { MainModule } from './main/main.module';
 import { MultiStepCheckoutPageLayoutModule } from './multi-step-checkout-page-layout/multi-step-checkout-page-layout.module';
 import { OrderConfirmationPageLayoutModule } from './order-confirmation-page-layout/order-confirmation-page-layout.module';
 import { OrderDetailsPageLayoutModule } from './order-details-page-layout/order-details-page-layout.module';
-import { AddressBookPageLayoutModule } from './address-book-page-layout/address-book-page-layout.module';
 import { OrderHistoryPageLayoutModule } from './order-history-page-layout/order-history-page-layout.module';
 import { PaymentDetailsPageLayoutModule } from './payment-details-page-layout/payment-details-page-layout.module';
 import { RegisterLayoutModule } from './register-layout/register-layout.module';
@@ -24,7 +23,6 @@ const layoutModules = [
   MultiStepCheckoutPageLayoutModule,
   OrderDetailsPageLayoutModule,
   OrderConfirmationPageLayoutModule,
-  AddressBookPageLayoutModule,
   PaymentDetailsPageLayoutModule,
   RegisterLayoutModule,
   LoginPageLayoutModule,
