--- projects/storefrontlib/src/lib/ui/pages/address-book-page/address-book-page.component.html
@@ -1,1 +1,0 @@
-<cx-address-book-page-layout></cx-address-book-page-layout>
