--- projects/storefrontlib/src/lib/ui/pages/address-book-page/address-book-page.component.ts
@@ -1,12 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'cx-address-book-page',
-  templateUrl: './address-book-page.component.html',
-  styleUrls: []
-})
-export class AddressBookPageComponent implements OnInit {
-  constructor() {}
-
-  ngOnInit() {}
-}
