+++ projects/storefrontlib/src/lib/ui/pages/myaccount/address-book-page/address-book-page.component.html
@@ -1,0 +1,6 @@
+<cx-page-layout>
+  <!-- 
+    We hard code the address book as long as we do not have a CMS component for this.
+  -->
+  <cx-address-book></cx-address-book>
+</cx-page-layout>
