+++ projects/storefrontlib/src/lib/ui/pages/myaccount/address-book-page/address-book-page.component.ts
@@ -1,0 +1,6 @@
+import { Component } from '@angular/core';
+
+@Component({
+  templateUrl: './address-book-page.component.html'
+})
+export class AddressBookPageComponent {}
